sched: Replace timezone dependent functions with UTC based functions.
Changed the usage of localtime and mktime to gmtime and timegm respectively.
This addresses inconsistencies related to DST and other timezone related
functionality on the Unix port with localtime and mktime, compared to the
other ports, where localtime and mktime do not use DST and directly report
the RCT time. By using gmtime and timegm these builds function the same.

For timegm, a custom implementation was added into the cron module, which also
required the need for the days_in_month and is_leap_year functions. Uses of the
_mdays dict has been replaced to use days_in_month which handles leap years.